# src/infrastructure/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateDocumentRequest": {
        "properties": {
          "audience": {
            "default": "caregiver",
            "enum": [
              "caregiver",
              "care_team",
              "child",
              "public"
            ],
            "title": "Audience",
            "type": "string"
          },
          "category": {
            "maxLength": 200,
            "title": "Category",
            "type": "string"
          },
          "language": {
            "default": "es",
            "maxLength": 8,
            "title": "Language",
            "type": "string"
          },
          "slug": {
            "maxLength": 120,
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title",
          "category"
        ],
        "title": "CreateDocumentRequest",
        "type": "object"
      },
...
      "CreateVersionRequest": {
        "properties": {
          "content": {
            "maxLength": 200000,
            "title": "Content",
            "type": "string"
          },
          "mimeType": {
            "default": "text/markdown",
            "maxLength": 120,
            "title": "Mimetype",
            "type": "string"
          },
          "version": {
            "maxLength": 32,
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "content"
        ],
        "title": "CreateVersionRequest",
        "type": "object"
      },
      "DocumentOut": {
        "properties": {
          "audience": {
            "title": "Audience",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isActive": {
            "title": "Isactive",
            "type": "boolean"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "category",
          "audience",
          "language",
          "isActive"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotificationOut": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "patientId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patientid"
          },
          "readAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readat"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body",
          "createdAt"
        ],
        "title": "NotificationOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VersionOut": {
        "properties": {
          "checksum": {
            "title": "Checksum",
            "type": "string"
          },
          "documentId": {
            "format": "uuid",
            "title": "Documentid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "publishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publishedat"
          },
          "retiredAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retiredat"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "id",
          "documentId",
          "version",
          "status",
          "checksum"
        ],
        "title": "VersionOut",
        "type": "object"
      },
      "VersionPreviewOut": {
        "properties": {
          "chunkCount": {
            "title": "Chunkcount",
            "type": "integer"
          },
          "sections": {
            "items": {
              "type": "string"
            },
            "title": "Sections",
            "type": "array"
          },
          "version": {
            "$ref": "#/components/schemas/VersionOut"
          }
        },
        "required": [
          "version",
          "chunkCount"
        ],
        "title": "VersionPreviewOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del piloto conectado de Kinti (Fase 2). Datos exclusivamente sint�ticos. El sem�foro representa riesgo operativo de interrupci�n de la ruta, no gravedad cl�nica. No diagnostica, no prescribe y no realiza triaje.",
    "title": "Kinti API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/documents": {
      "get": {
        "operationId": "list_documents_api_v1_knowledge_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentOut"
                  },
                  "title": "Response List Documents Api V1 Knowledge Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "conocimiento"
        ]
      },
      "post": {
        "operationId": "create_document_api_v1_knowledge_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Document",
        "tags": [
          "conocimiento"
        ]
      }
    },
    "/api/v1/knowledge/documents/{document_id}/versions": {
      "post": {
        "operationId": "create_version_api_v1_knowledge_documents__document_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Version",
        "tags": [
          "conocimiento"
        ]
      }
    },
    "/api/v1/knowledge/versions/{version_id}/preview": {
      "get": {
        "operationId": "preview_version_api_v1_knowledge_versions__version_id__preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionPreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Preview Version",
        "tags": [
          "conocimiento"
        ]
      }
    },
    "/api/v1/knowledge/versions/{version_id}/process": {
      "post": {
        "description": "Extrae, fragmenta y embebe. Deja la versi�n pendiente de revisi�n.",
        "operationId": "process_version_api_v1_knowledge_versions__version_id__process_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Process Version",
        "tags": [
          "conocimiento"
        ]
      }
    },
    "/api/v1/knowledge/versions/{version_id}/publish": {
      "post": {
        "operationId": "publish_version_api_v1_knowledge_versions__version_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Version",
        "tags": [
          "conocimiento"
        ]
      }
    },
    "/api/v1/knowledge/versions/{version_id}/retire": {
      "post": {
        "operationId": "retire_version_api_v1_knowledge_versions__version_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retire Version",
        "tags": [
          "conocimiento"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "list_notifications_api_v1_notifications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationOut"
                  },
                  "title": "Response List Notifications Api V1 Notifications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Notifications",
        "tags": [
          "notificaciones"
        ]
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "operationId": "mark_read_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark Read",
        "tags": [
          "notificaciones"
        ]
      }
    },
...
  }
}